+++ package.json
@@ -98,6 +98,7 @@
     "rollup": "^1.32.0",
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-terser": "^5.2.0",
+    "rollup-plugin-web-worker-loader": "^0.9.0",
     "strict-uri-encode": "^2.0.0",
     "twitter-text": "^3.0.1"
   },
