+++ src/pages/admin/index.pug
@@ -7,5 +7,4 @@ html(lang="ja")
     title 管理画面（予定）
     meta(name="robots" content="noindex")
   body
-    script(src="/asciidoctor/asciidoctor.min.js" defer)
-    script(type="module" src="/admin/main.js")
+    script(type="module" src="/admin/main.js" async)
